Patch to avoid unexpected behaviour when no seed is set.
This patch affects the functions ugarchpath and ugarchsim when the
parameter rseed is not set (i. e., when it has its default value NA).

The previous behaviour was that an integer chosen randomly according to
a uniform distribution was chosen between 0 and Sys.time().

The new behaviour is that, in case rseed == NA, the seed of the RNG of R
is left unmodified to not interfer with the process of generating
numbers that can be considered to be independent.